+++ KittsInfractionSystem/Properties/AssemblyInfo.cs
@@ -25,8 +25,7 @@ using System.Runtime.InteropServices;
 //
 //      Major Version
 //      Minor Version
-//      Build Number
-//      Revision
+//      Fix Version
 //
-[assembly: AssemblyVersion("0.1.1")]
+[assembly: AssemblyVersion("0.2.0")]
 [assembly: AssemblyFileVersion("1.0.0.0")]