+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # KittsInfractionSystem Changelog
 
+## Version 0.2.0
+- Added `GetPrettyInfraction` function - Gets a pretty string for an `InfractionData`.
+- Added `GetPrettyInfractions` function - Gets a pretty string from a list of `InfractionData` or by an offender's id.
+- Added `GetPrettyColouredInfraction` function - Gets a pretty coloured string for an `InfractionData`.
+- Added `GetPrettyColouredInfractions` function - Gets a pretty coloured string from a list of `InfractionData` or by an offender's id.
+- Updated `FormatDuration` function - Now returns time down to the second.
+- Updated `KittsInfractionSystem` - Now split into a normal and MongoDB version.
+
 ## Version 0.1.1
 - Updated `KittsInfractionSystem` class - Now `public` instead of `internal`, rookie mistake.
 - Updated `Dependencies` - Removed unnecessary dependencies.